feat(gold-reward): implement gold reward processing and enhance event management features
- Marked the implementation of the gold reward distribution system as complete in the event management tasks.
- Deleted obsolete permission creation commands for import/export functionalities.
- Introduced a new command to process failed gold rewards, enhancing the reliability of gold distribution.
- Updated the event completion command to include processing of failed gold rewards and improved logging.
- Enhanced the EventParticipationService with validation and logging for gold reward eligibility and auditing.

This update improves the event management system by ensuring accurate and efficient processing of gold rewards, enhancing user experience and operational reliability.

+++ app/Http/Resources/Api/V1/Lecturer/SessionResource.php
@@ -173,7 +173,7 @@ class SessionResource extends JsonResource
         $sessionDate = \Carbon\Carbon::parse($date);
 
         return in_array($status, ['completed', 'in_progress']) ||
-               ($status === 'scheduled' && $sessionDate->lte(now()));
+            ($status === 'scheduled' && $sessionDate->lte(now()));
     }
 
     /**
@@ -197,7 +197,7 @@ class SessionResource extends JsonResource
             return false;
         }
 
-        $sessionDateTime = \Carbon\Carbon::parse($date.' '.$startTime);
+        $sessionDateTime = \Carbon\Carbon::parse($date . ' ' . $startTime);
 
         return $sessionDateTime->isPast();
     }
@@ -211,7 +211,7 @@ class SessionResource extends JsonResource
             return false;
         }
 
-        $sessionDateTime = \Carbon\Carbon::parse($date.' '.$startTime);
+        $sessionDateTime = \Carbon\Carbon::parse($date . ' ' . $startTime);
 
         return $sessionDateTime->isFuture();
     }
@@ -225,13 +225,13 @@ class SessionResource extends JsonResource
             return null;
         }
 
-        $sessionDateTime = \Carbon\Carbon::parse($date.' '.$startTime);
+        $sessionDateTime = \Carbon\Carbon::parse($date . ' ' . $startTime);
 
         if ($sessionDateTime->isPast()) {
             return null;
         }
 
-        return now()->diffInMinutes($sessionDateTime);
+        return (int) now()->diffInMinutes($sessionDateTime);
     }
 
     /**
@@ -261,7 +261,7 @@ class SessionResource extends JsonResource
      */
     protected function getFormattedTime(string $startTime, string $endTime): string
     {
-        return $startTime.' - '.$endTime;
+        return $startTime . ' - ' . $endTime;
     }
 
     /**
@@ -273,7 +273,7 @@ class SessionResource extends JsonResource
             return null;
         }
 
-        $sessionDateTime = \Carbon\Carbon::parse($date.' '.$startTime);
+        $sessionDateTime = \Carbon\Carbon::parse($date . ' ' . $startTime);
 
         return $sessionDateTime->diffForHumans();
     }
